Static method created
